switched to using fixed autosummary everywhere
